# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus AI Platform - AI Recruitment & CFO SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates
      parameters:
        - name: job_id
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Add a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /candidates/{id}/advance:
    post:
      operationId: advanceCandidate
      tags: [candidates]
      summary: Advance candidate to next pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageAdvance"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}/score:
    post:
      operationId: scoreCandidate
      tags: [candidates]
      summary: AI score a candidate against job requirements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiScorecard"

  # INTERVIEWS
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Onboard a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Offboard employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Candidate:
      type: object
      required: [id, full_name, email, job_id, stage, ai_score, created_at]
      properties:
        id:
          type: string
        full_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        job_id:
          type: string
        job_title:
          type: ["string", "null"]
        stage:
          type: string
          enum: [applied, screening, interview, offer, hired, rejected]
        ai_score:
          type: ["number", "null"]
        skills:
          type: array
          items:
            type: string
        experience_years:
          type: ["number", "null"]
        current_company:
          type: ["string", "null"]
        current_title:
          type: ["string", "null"]
        expected_salary:
          type: ["number", "null"]
        resume_url:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    CandidateInput:
      type: object
      required: [full_name, email, job_id]
      properties:
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        job_id:
          type: string
        stage:
          type: string
        skills:
          type: array
          items:
            type: string
        experience_years:
          type: number
        current_company:
          type: string
        current_title:
          type: string
        expected_salary:
          type: number
        resume_url:
          type: string
        notes:
          type: string
        source:
          type: string
    CandidateUpdate:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        job_id:
          type: string
        stage:
          type: string
        skills:
          type: array
          items:
            type: string
        experience_years:
          type: number
        current_company:
          type: string
        current_title:
          type: string
        expected_salary:
          type: number
        resume_url:
          type: string
        notes:
          type: string
        source:
          type: string
    StageAdvance:
      type: object
      required: [stage]
      properties:
        stage:
          type: string
          enum: [applied, screening, interview, offer, hired, rejected]
        notes:
          type: string
    AiScorecard:
      type: object
      required: [overall_score, skills_match, experience_match, recommendation]
      properties:
        overall_score:
          type: number
        skills_match:
          type: number
        experience_match:
          type: number
        cultural_fit:
          type: number
        recommendation:
          type: string
          enum: [strong_hire, hire, maybe, no_hire]
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        risk_factors:
          type: array
          items:
            type: string
        suggested_salary:
          type: ["number", "null"]

    # INTERVIEWS
...
    Employee:
      type: object
      required: [id, full_name, email, department, job_title, employment_type, status, start_date, base_salary, created_at]
      properties:
        id:
          type: string
        full_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: string
        job_title:
          type: string
        employment_type:
          type: string
          enum: [full_time, part_time, contract]
        status:
          type: string
          enum: [active, on_leave, terminated]
        start_date:
          type: string
        end_date:
          type: ["string", "null"]
        base_salary:
          type: number
        performance_score:
          type: ["number", "null"]
        manager_id:
          type: ["string", "null"]
        manager_name:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        created_at:
          type: string
    EmployeeInput:
      type: object
      required: [full_name, email, department, job_title, employment_type, start_date, base_salary]
      properties:
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        job_title:
          type: string
        employment_type:
          type: string
        start_date:
          type: string
        base_salary:
          type: number
        manager_id:
          type: string
        location:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        job_title:
          type: string
        employment_type:
          type: string
        status:
          type: string
        base_salary:
          type: number
        performance_score:
          type: number
        manager_id:
          type: string
        location:
          type: string
        end_date:
          type: string
...