# lib/api-zod/src/generated/types/aiScorecard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus AI Platform - AI Recruitment & CFO SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { AiScorecardRecommendation } from './aiScorecardRecommendation';

export interface AiScorecard {
  overall_score: number;
  skills_match: number;
  experience_match: number;
  cultural_fit?: number;
  recommendation: AiScorecardRecommendation;
  strengths?: string[];
  weaknesses?: string[];
  risk_factors?: string[];
  /** @nullable */
  suggested_salary?: number | null;
}